test: move test for new component
