+++ frontend/src/tests/lib/pages/NnsNeurons.spec.ts
@@ -3,22 +3,19 @@
  */
 
 import { NeuronState } from "@dfinity/nns";
-import { fireEvent, render, waitFor } from "@testing-library/svelte";
+import { render, waitFor } from "@testing-library/svelte";
 import NnsNeurons from "../../../lib/pages/NnsNeurons.svelte";
 import { authStore } from "../../../lib/stores/auth.store";
 import { neuronsStore } from "../../../lib/stores/neurons.store";
-import { voteInProgressStore } from "../../../lib/stores/voting.store";
 import {
   mockAuthStoreSubscribe,
   mockPrincipal,
 } from "../../mocks/auth.store.mock";
-import en from "../../mocks/i18n.mock";
 import {
   buildMockNeuronsStoreSubscribe,
   mockFullNeuron,
   mockNeuron,
 } from "../../mocks/neurons.mock";
-import { mockVotingInProgressItem } from "../../mocks/proposal.mock";
 
 jest.mock("../../../lib/services/neurons.services", () => {
   return {
@@ -101,59 +98,6 @@ describe("NnsNeurons", () => {
       waitFor(() =>
         expect(container.querySelector('article[role="link"]')).not.toBeNull()
       );
-    });
-
-    it("should open the CreateNeuronModal on click to Stake Neurons", async () => {
-      const { queryByTestId, queryByText } = render(NnsNeurons);
-
-      const toolbarButton = queryByTestId("stake-neuron-button");
-      expect(toolbarButton).not.toBeNull();
-      expect(queryByText(en.accounts.select_source)).toBeNull();
-
-      toolbarButton !== null && (await fireEvent.click(toolbarButton));
-
-      expect(queryByText(en.accounts.select_source)).not.toBeNull();
-    });
-
-    it("should disable Stake Neurons button during voting process", async () => {
-      const { queryByTestId } = render(NnsNeurons);
-
-      const stakeNeuronButton = queryByTestId("stake-neuron-button");
-      expect(stakeNeuronButton).not.toBeNull();
-
-      voteInProgressStore.add(mockVotingInProgressItem);
-
-      waitFor(() =>
-        expect(stakeNeuronButton?.getAttribute("disabled")).not.toBeNull()
-      );
-    });
-
-    it("should open the MergeNeuronsModal on click to Merge Neurons", async () => {
-      const { queryByTestId, queryByText } = render(NnsNeurons);
-
-      const toolbarButton = queryByTestId("merge-neurons-button");
-      expect(toolbarButton).not.toBeNull();
-      expect(queryByText(en.neurons.merge_neurons_modal_title)).toBeNull();
-
-      toolbarButton !== null && (await fireEvent.click(toolbarButton));
-
-      expect(queryByText(en.neurons.merge_neurons_modal_title)).not.toBeNull();
-    });
-  });
-
-  describe("with less than two neurons", () => {
-    beforeEach(() => {
-      jest
-        .spyOn(neuronsStore, "subscribe")
-        .mockImplementation(buildMockNeuronsStoreSubscribe([mockNeuron]));
-    });
-    it("should have disabled Merge Neurons button", async () => {
-      const { queryByTestId } = render(NnsNeurons);
-
-      const toolbarButton = queryByTestId("merge-neurons-button");
-      expect(toolbarButton).not.toBeNull();
-      toolbarButton &&
-        expect(toolbarButton.hasAttribute("disabled")).toBeTruthy();
     });
   });
 });
